Added Authorization classes

+++ backend/src/ServiceReleaseManager.Api/Endpoints/OrganisationRoles/Delete.cs
@@ -1,6 +1,8 @@
 ﻿using Ardalis.Result.AspNetCore;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using ServiceReleaseManager.Api.Authorization;
+using ServiceReleaseManager.Api.Authorization.Operations.Organisation;
 using ServiceReleaseManager.Core.Interfaces;
 using Swashbuckle.AspNetCore.Annotations;
 
@@ -9,10 +11,12 @@ namespace ServiceReleaseManager.Api.Endpoints.OrganisationRoles;
 public class Delete : EndpointBase.WithRequest<DeleteOrganisationRoleRequest>.WithoutResult
 {
   private readonly IOrganisationRoleService _organisationRoleService;
+  private readonly IServiceManagerAuthorizationService _authorizationService;
 
-  public Delete(IOrganisationRoleService organisationRoleService)
+  public Delete(IOrganisationRoleService organisationRoleService, IServiceManagerAuthorizationService authorizationService)
   {
     _organisationRoleService = organisationRoleService;
+    _authorizationService = authorizationService;
   }
 
   [HttpDelete(DeleteOrganisationRoleRequest.Route)]
@@ -26,6 +30,12 @@ public class Delete : EndpointBase.WithRequest<DeleteOrganisationRoleRequest>.Wi
     [FromRoute] DeleteOrganisationRoleRequest request,
     CancellationToken cancellationToken = new())
   {
+    if (!await _authorizationService.EvaluateOrganisationAuthorization(User,
+      OrganisationRoleOperation.OrganisationRole_Delete, cancellationToken))
+    {
+      return NotFound();
+    }
+
     var result =
       await _organisationRoleService.Delete(request.OrganisationRoleId, cancellationToken);
