+++ backend/src/ServiceReleaseManager.Api/Endpoints/OrganisationUsers/Delete.cs
@@ -1,5 +1,7 @@
 ﻿using Ardalis.Result.AspNetCore;
 using Microsoft.AspNetCore.Mvc;
+using ServiceReleaseManager.Api.Authorization;
+using ServiceReleaseManager.Api.Authorization.Operations.Organisation;
 using ServiceReleaseManager.Core.Interfaces;
 using Swashbuckle.AspNetCore.Annotations;
 
@@ -8,10 +10,12 @@ namespace ServiceReleaseManager.Api.Endpoints.OrganisationUsers;
 public class Delete : EndpointBase.WithRequest<DeleteOrganisationUserRequest>.WithoutResult
 {
   private readonly IOrganisationUserService _organisationUserService;
+  private readonly IServiceManagerAuthorizationService _authorizationService;
 
-  public Delete(IOrganisationUserService organisationUserService)
+  public Delete(IOrganisationUserService organisationUserService, IServiceManagerAuthorizationService authorizationService)
   {
     _organisationUserService = organisationUserService;
+    _authorizationService = authorizationService;
   }
 
   [HttpDelete(DeleteOrganisationUserRequest.Route)]
@@ -25,6 +29,12 @@ public class Delete : EndpointBase.WithRequest<DeleteOrganisationUserRequest>.Wi
     [FromRoute] DeleteOrganisationUserRequest request,
     CancellationToken cancellationToken = new())
   {
+    if(!await _authorizationService.EvaluateOrganisationAuthorization(User,
+      OrganisationUserOperations.OrganisationUser_Delete, cancellationToken))
+    {
+      return NotFound();
+    }
+
     var result =
       await _organisationUserService.Delete(request.OrganisationUserId, cancellationToken);
